Fixes after PR review
