+++ src/oned/mod.rs
@@ -93,4 +93,3 @@ mod upc_e_writer;
 pub use upc_e_writer::*;
 
 mod telepen_common;
-pub use telepen_common::*;
